app/api/serializer.py
from rest_framework import serializers
from app.models import Manufacturer


class ManufacturerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Manufacturer
        fields = ['name', 'created_at']